+++ src/main/java/com/github/jeffmagina/restaurant/menu/Menu.java
@@ -4,43 +4,45 @@ import com.github.jeffmagina.restaurant.food.Food;
 import java.util.ArrayList;
 
 public class Menu {
-    
-    //Create menu storage
-    private ArrayList<Food> menu = new ArrayList<Food>();
-    
-    //*********************Make menu maybe map so i can count how many steaks for example and display that */
-    
-    public Menu(){
-    	populateMenu();
-    }
-    
-    public void populateMenu(){
-        addMenuItem("chicken", 9.50);
-        addMenuItem("steak", 10.50);
-        addMenuItem("fish", 8.50);
-        addMenuItem("vegetarian", 6.00);
-        addMenuItem("fries", 3.00); 
-    }
 
-    public void addMenuItem(String name, double cost){
-        Food menuItem = new Food(name,cost);
-         this.menu.add(menuItem);
-    }
+	// Create menu storage
+	private ArrayList<Food> menu = new ArrayList<Food>();
 
-    public void displayMenu(){
-        for(int i = 0; i < this.menu.size(); i++){
-            System.out.println(this.menu.get(i).name + ", " + this.menu.get(i).cost);  
-        }
-    }
+	// *********************Make menu maybe map so i can count how many steaks for
+	// example and display that */
 
-    public int size(){
-        return this.menu.size();
-    }
+	public Menu() {
+		populateMenu();
+	}
 
-    public Food getItem(int i){
-        if (i >= this.menu.size() || i < 0){
-            throw new IndexOutOfBoundsException("Index: " + i + ", Size " + i);
-        } return this.menu.get(i);
-    }
+	public void populateMenu() {
+		addMenuItem("Chicken", 9.50);
+		addMenuItem("Steak", 10.50);
+		addMenuItem("Fish", 8.50);
+		addMenuItem("Vegetarian", 6.00);
+		addMenuItem("Fries", 3.00);
+	}
+
+	public void addMenuItem(String name, double cost) {
+		Food menuItem = new Food(name, cost);
+		this.menu.add(menuItem);
+	}
+
+	public void displayMenu() {
+		for (int i = 0; i < this.menu.size(); i++) {
+			System.out.println(this.menu.get(i).name + ", " + this.menu.get(i).cost);
+		}
+	}
+
+	public int size() {
+		return this.menu.size();
+	}
+
+	public Food getItem(int i) {
+		if (i >= this.menu.size() || i < 0) {
+			throw new IndexOutOfBoundsException("Index: " + i + ", Size " + i);
+		}
+		return this.menu.get(i);
+	}
 
 }