+++ src/main/java/com/github/jeffmagina/restaurant/orderhistory/OrderHistory.java
@@ -1,10 +1,24 @@
 package com.github.jeffmagina.restaurant.orderhistory;
 
-//import java.util.ArrayList;
 import com.github.jeffmagina.restaurant.customerorder.CustomerOrder;
+import java.util.ArrayList;
+//import com.github.jeffmagina.restaurant.customerorder.CustomerOrder;
 
 public class OrderHistory extends CustomerOrder {
 
-    public OrderHistory(){}
+	static ArrayList<CustomerOrder> allOrders = new ArrayList<CustomerOrder>();
 
+	public OrderHistory() {
+	}
+
+	public void storeOrder(CustomerOrder custOrder) {
+		this.allOrders.add(custOrder);
+		// store in database
+	}
+
+	public void displayAllOrders() {
+    	for(int i = 0; i < allOrders.size(); i++) {
+    	System.out.println(this.allOrders.get(i));
+    	}
+    }
 }